+++ packages/markdown/index.js
@@ -1,3 +1,4 @@
 export { default as mermaid } from 'mermaid';
+export { default as katex } from 'katex';
 export { formatMarkdown, formatHtmlToMarkdown, markdownParser, markdownToAnnotatedText, renderMarkdownToHtml } from './mdext/index.js';